refactor(api): rilocate slack_api_connection file in package
* create package api and rilocate slack_api_connection file
* update import
* update tests

Closes #22
