+++ tests/api_test/slack_api_connection_integration_test.py
@@ -1,5 +1,6 @@
 from chatbot import cadocs
-from src import slack_api_connection, service, chatbot
+from src import api, service, chatbot
+from api import slack_api_connection
 from service import utils
 from chatbot.cadocs import Cadocs
 from intent_handling.cadocs_intents import CadocsIntents
@@ -43,7 +44,7 @@ class TestSlackAPIConnectionIT:
         mocker.patch.object(slack_api_connection.WebClient,
                             "chat_postMessage", return_value=data_test.get("message"))
 
-        mocker.patch('slack_api_connection.os.environ.get', side_effect=[
+        mocker.patch('api.slack_api_connection.os.environ.get', side_effect=[
             "CADOCSNLU_URL_PREDICT", "0.77", "0.55", "0.77", "CSDETECTOR_URL_GETSMELLS"])
 
         response_intent_manager = {
@@ -71,7 +72,7 @@ class TestSlackAPIConnectionIT:
         mock_response_tools.json.return_value = response
 
         # Mock requests.get method
-        mocker.patch("slack_api_connection.requests.get",
+        mocker.patch("api.slack_api_connection.requests.get",
                      side_effect=[mock_response_intent_manager, mock_response_tools])
 
         # Mock os.environ.get method
