+++ tests/api_test/slack_api_connection_unit_test.py
@@ -1,8 +1,9 @@
-from src import slack_api_connection, service
+from src import api, service
+from api import slack_api_connection
 from src.service import utils
 from intent_handling.cadocs_intents import CadocsIntents
 from flask import Flask, make_response
-from slack_api_connection import app, Cadocs
+from api.slack_api_connection import app, Cadocs
 import json
 import pytest
 import requests
@@ -66,7 +67,7 @@ class TestSlackAPIConnectionUT:
         # Mock the post_waiting_message function
         mock_post_waiting_message = Mock()
         mock_post_waiting_message.do_run = True
-        mocker.patch('src.slack_api_connection.post_waiting_message',
+        mocker.patch('api.slack_api_connection.post_waiting_message',
                      return_value=mock_post_waiting_message)
         # Mock the new_message function of Cadocs class
         mocker.patch.object(slack_api_connection.Cadocs, "new_message", return_value=(
@@ -106,7 +107,7 @@ class TestSlackAPIConnectionUT:
         # Mock the post_waiting_message function
         mock_progress_message = Mock()
         mock_progress_message.do_run = True
-        mocker.patch('src.slack_api_connection.post_waiting_message',
+        mocker.patch('api.slack_api_connection.post_waiting_message',
                      return_value=mock_progress_message)
         # Mock the new_message function of Cadocs class to make it throw an exception
         mock_new_message = Mock(side_effect=Exception("test exception"))
@@ -130,7 +131,7 @@ class TestSlackAPIConnectionUT:
     def test_get_smells_wrepo_wpat(self, mocker, client, args, expected_url):
         # Mock of the Response object
         mock_response = mocker.Mock(spec=requests.Response)
-        mocker.patch("slack_api_connection.requests.get",
+        mocker.patch("api.slack_api_connection.requests.get",
                      return_value=mock_response)
 
         client.get(f"/csDetector/getSmells{args}")
@@ -146,7 +147,7 @@ class TestSlackAPIConnectionUT:
     def test_get_smells_no_repo_no_pat(self, mocker, client, expected_response, args):
         # Mock of the Response object
         mock_response = mocker.Mock(spec=requests.Response)
-        mocker.patch("slack_api_connection.requests.get",
+        mocker.patch("api.slack_api_connection.requests.get",
                      return_value=mock_response)
 
         response = client.get(f"/csDetector/getSmells{args}")
@@ -163,7 +164,7 @@ class TestSlackAPIConnectionUT:
         # Mock of the Response object
         mock_response = mocker.Mock(spec=requests.Response)
         mock_response.content = "test_content"
-        mocker.patch("slack_api_connection.requests.get",
+        mocker.patch("api.slack_api_connection.requests.get",
                      return_value=mock_response)
 
         response = client.get(
@@ -177,7 +178,7 @@ class TestSlackAPIConnectionUT:
     def test_predict(self, mocker, client):
         # Mock of the Response object
         mock_response = mocker.Mock(spec=requests.Response)
-        mocker.patch("slack_api_connection.requests.get",
+        mocker.patch("api.slack_api_connection.requests.get",
                      return_value=mock_response)
 
         client.get("/cadocsNLU/predict?message=hello i'm tester")
@@ -190,7 +191,7 @@ class TestSlackAPIConnectionUT:
     def test_predict_no_message(self, mocker, client):
         # Mock of the Response object
         mock_response = mocker.Mock(spec=requests.Response)
-        mocker.patch("slack_api_connection.requests.get",
+        mocker.patch("api.slack_api_connection.requests.get",
                      return_value=mock_response)
 
         response = client.get("/cadocsNLU/predict")
@@ -202,7 +203,7 @@ class TestSlackAPIConnectionUT:
     def test_update_dataset(self, mocker, client):
         # Mock of the Response object
         mock_response = mocker.Mock(spec=requests.Response)
-        mocker.patch("slack_api_connection.requests.get",
+        mocker.patch("api.slack_api_connection.requests.get",
                      return_value=mock_response)
 
         client.get(
@@ -217,7 +218,7 @@ class TestSlackAPIConnectionUT:
     def test_update_dataset_no_message_no_intent(self, mocker, client, args):
         # Mock of the Response object
         mock_response = mocker.Mock(spec=requests.Response)
-        mocker.patch("slack_api_connection.requests.get",
+        mocker.patch("api.slack_api_connection.requests.get",
                      return_value=mock_response)
 
         response = client.get(f"/cadocsNLU/update{args}")
